pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elfomofi-sdk"
version = "0.1.0"
description = "Python SDK for ElfomoFi — onchain quoting with offchain interpolation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["defi", "amm", "quoting", "orderbook", "web3", "ethereum", "base"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "web3>=6.0,<8.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
]
examples = [
    "python-dotenv>=1.0",
]

[project.urls]
Repository = "https://github.com/ElfomoFi/elfomofi-sdk"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
include = ["elfomofi_sdk*"]

[tool.setuptools.package-data]
elfomofi_sdk = ["abi/*.json", "py.typed"]
